### 2020-06-11 #2319 @whedon
[PRE REVIEW]: shinyssdtools: A web application for fitting Species Sensitivity Distributions (SSDs)
**Submitting author:** @sebdalgarno (<a href="http://orcid.org/0000-0002-3658-4517">Sebastian Dalgarno</a>)
**Repository:** <a href="https://github.com/bcgov/shinyssdtools" target ="_blank">https://github.com/bcgov/shinyssdtools</a>
**Version:** v0.0.1
