### 2020-06-11 #2319 @arfon
This is because the YAML header is malformed. @sebdalgarno - please update the YAML header of your `paper.md` to look like this:

```
---
title: 'shinyssdtools: A web application for fitting Species Sensitivity Distributions (SSDs)'
date: '2020-06-02'
authors:
- affiliation: 1
  name: Seb Dalgarno
  orcid: 0000-0002-3658-4517
affiliations:
- name: Poisson Consulting, Nelson, British Columbia
  index: 1
bibliography: paper.bib
tags:
   - R
   - Shiny
   - ssdtools
   - species sensitivity distributions
---
```
