+++ README.md
@@ -2,7 +2,8 @@
 
 <a id="devex-badge" rel="Delivery" href="https://github.com/BCDevExchange/assets/blob/master/README.md"><img alt="In production, but maybe in Alpha or Beta. Intended to persist and be supported." style="border-width:0" src="https://assets.bcdevexchange.org/images/badges/delivery.svg" title="In production, but maybe in Alpha or Beta. Intended to persist and be supported." /></a>
 
-An R package to run a Shiny interface to the [ssdtools](https://cran.r-project.org/package=ssdtools) R package.
+An R package to run a Shiny interface to the [ssdtools](https://cran.r-project.org/package=ssdtools) R package. 
+The app can be run locally or by visiting [this site](https://bcgov-env.shinyapps.io/ssdtools/).
 
 ### Getting Help or Reporting an Issue
 
