### 2020-06-18 #2319 @Kevin-Mattheus-Moerman
@sebdalgarno thank you for your submission. The editorial board has decided this work is unfortunately not in scope for JOSS (see also our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and statements on "Minor ‘utility’ packages, including ‘thin’ API clients, and single-function packages").  Of course this conclusion says nothing about the value/usefulness of the work submitted, it merely means JOSS is not the right place to publish it. We will therefore proceed to reject this work as it stands.
