+++ README.md
@@ -1,9 +1,17 @@
-# ssdtools-shiny
+# shinyssdtools
 
 <a id="devex-badge" rel="Delivery" href="https://github.com/BCDevExchange/assets/blob/master/README.md"><img alt="In production, but maybe in Alpha or Beta. Intended to persist and be supported." style="border-width:0" src="https://assets.bcdevexchange.org/images/badges/delivery.svg" title="In production, but maybe in Alpha or Beta. Intended to persist and be supported." /></a>
 
-An R package to run a Shiny interface to the [ssdtools](https://cran.r-project.org/package=ssdtools) R package. 
-The app can be run locally or by visiting [this site](https://bcgov-env.shinyapps.io/ssdtools/).
+A Shiny web application for fitting Species Sensitivity Distributions (SSDs). 
+
+### Summary
+`shinyssdtools` provides a Graphical User Interface (GUI) to the [ssdtools](https://cran.r-project.org/package=ssdtools) R package. The app can be run locally by installing the R package and running `shinyssdtools::run_ssdtools_app()` or by visiting [this site](https://bcgov-env.shinyapps.io/ssdtools/).
+
+In addition to providing access to core functionality in the [ssdtools](https://cran.r-project.org/package=ssdtools) R package, `shinyssdtools` provides:  
+
+1. translation to French 
+2. generation of R code required to reproduce results
+3. further customization of plot and table outputs
 
 ### Getting Help or Reporting an Issue
 
