### 2020-06-18 #2319 @sebdalgarno
Hi @Kevin-Mattheus-Moerman I've just updated our README to indicate the additional functionality and value that this software adds to ssdtools. This includes: translation to French; generation of R code for user to reproduce results; further customization of plot and table outputs.
