### 2020-06-18 #2319 @joethorley
Hi @Kevin-Mattheus-Moerman as a reviewer for JOSS and  the co-author of the underlying ssdtools R package I'd ask that you reconsider your initial decision based on the additional information provided by @sebdalgarno. shinyssdtools is an R package of over 1,300 lines of code and 14 functions that took @sebdalgarno than 110 hours to program (I'm also his line-manager!). In addition to providing a multi-panel user-friendly GUI, and adding the functionality indicated above, it also provides substantially more features than the shinyssd R package that was recently published in JOSS  ([D'Andrea et al. 2019](https://joss.theoj.org/papers/10.21105/joss.00785)). shinyssdtools was commissioned by the BC/Canadian government, is being considered for adoption by the Australian government and represents an important development for many users in the field. I recommended to @sebdalgarno that he attempt to publish in JOSS as I am a co-author on a paper that is ready for submission with 8 other Canadian/Australian scientists that reviews the software available for fitting SSDs. I am hoping that we will be able to cite shinyssdtools so that @sebdalgarno receives full credit for his work.
