### 2020-07-15 #2319 @Kevin-Mattheus-Moerman
@joethorley @sebdalgarno apologies for the delay in getting back to you. I will consider your comments and will bring this up with the board soon.
You say you expanded the readme with notes on functionality. Can you create/reproduce a clear description of the key functionality here and also work it into the paper? Can you make sure you include a clear description of the significant impact this software has on scientific research? This would help us when we reconsider this work.
I'll recreate the paper here now too and you can run `@whedon generate pdf` to update the paper if you like.
