### 2020-07-15 971abad
update paper for reconsideration
