+++ paper/paper.md
@@ -18,7 +18,7 @@ tags:
 
 # Summary
 
-`shinyssdtools` is a Shiny [@shiny] web application and R package for fitting Species Sensitivity Distributions (SSDs). It has been adopted by the governments of British Columbia, Canada and Australia to set water quality benchmarks and to characterize the effects of chemical contaminants for water quality and/or ecological risk assessment purposes. 
+`shinyssdtools` is a Shiny [@shiny] web application and R package for fitting Species Sensitivity Distributions (SSDs). It was developed for the Province of British Columbia with input from the governments of Canada as well as Australia and New Zealand. The shinyssdtools app has been used by the governments of B.C. and Canada to derive water quality benchmarks.
 
 SSDs are used to estimate the concentration of a chemical that affects a certain percentage (typically 5%) of the species considered [@posthuma_species_2001]. 
 The `ssdtools` R package [@ssdtools] allows model averaging using information-theoretic criteria and the construction of confidence intervals using bootstrapping [@ssdtools].
@@ -27,7 +27,7 @@ The `ssdtools` R package [@ssdtools] allows model averaging using information-th
 
 ### Graphical User Interface
 
-The `shinyssdtools` web application has five navigational tabs: 
+The `shinyssdtools` web application has six navigational tabs: 
 
 1. Data
    - Upload a dataset or enter data manually.
@@ -39,6 +39,8 @@ The `shinyssdtools` web application has five navigational tabs:
    - Calculate confidence limits using bootstrapping.
 1. R code
    - Copy the R code required to reproduce the results.
+1. About
+   - version information, explanation of abbreviations and references
 1. User guide
    - Step-by-step guide to proper use of the application.
 
@@ -58,11 +60,11 @@ shinyssdtools::run_ssdtools_app()
 # Contribution
 
 The similarly named `shinyssd` is an alternative open source Shiny web application to fit SSDs that is also bundled as an R package [@dandrea_shinyssd_2019].
-`shinyssdtools` contributes by being bilingual; providing additonal distributions including the gamma, Gompertz and log-Gumbel; by allowing the user to model average and by providing the R code to replicate the analysis.
+`shinyssdtools` contributes by being bilingual; providing additional distributions including the gamma, Gompertz and log-Gumbel; by allowing the user to model average and by providing the R code to replicate the analysis.
 
 # Acknowledgements
 
 We acknowledge contributions from Angeline Tillmanns, Marianne Métivier, Andy Teucher, David Fox, Carl Schwarz and Joe Thorley.
-Development of `shinyssdtools` was funded by the Ministry of Environment and Climate Change Strategy, British Columbia. The governments of British Columbia, Canada and Australia have also contributed to it's development.
+Development of `shinyssdtools` was funded by the Ministry of Environment and Climate Change Strategy, British Columbia. The governments of British Columbia, Canada and Australia and New Zealand have also contributed to its development.
 
 ## References
