### 2020-07-27 #2319 @sebdalgarno
Hi @Kevin-Mattheus-Moerman. Thanks for the suggestions to expand on functionality that shinyssdtools provides in addition to sddtools and it’s impact contribution to the scientific community. I’ve now added these to the paper and will now regenerate the pdf.

With respect to additional functionality, shinyssdtools provides: translation to French; generation of R code to reproduce results from a session; further customization of plot and table outputs; download of tables and plot outputs.

With respect to impact on scientific community: It has been adopted by the governments of British Columbia, Canada and Australia to set water quality benchmarks and to characterize the effects of chemical contaminants for water quality and/or ecological risk assessment purposes.

Finally, it’s worth noting that a manuscript which reviews Species Sensitivity Distribution software in Environmental Toxicology and Chemistry (Manuscript ID ETCJ-Jun-20-00450) has been accepted needing major revisions. If this shinyssdtools manuscript is accepted by August 7th, it will be cited in that manuscript.
