### 2020-08-08 12d44e4
update to CRAN ssdtools version
