### 2020-08-12 a7e6869
dynamically find ssdtools version in about
