+++ inst/extdata/about_fr.md
@@ -1,4 +1,4 @@
-Cette application web ajuste les fonctions de distribution de la sensibilité des espèces aux données de concentration. L'application est basée sur le progiciel R ssdtools version 0.2.0 et partage les mêmes fonctions. Les changements et les mises à niveau apportés à ssdtools vont résulter en modifications de cette application. Il est recommandé, lorsque l'estimation du HC5 généré par cette application est rapportée, d'inclure aussi la version de ssdtools et le nom des fonctions de distribution ajustées aux données.
+Cette application web ajuste les fonctions de distribution de la sensibilité des espèces aux données de concentration. L'application est basée sur le progiciel R ssdtools, et partage les mêmes fonctions. Les changements et les mises à niveau apportés à ssdtools vont résulter en modifications de cette application. Il est recommandé, lorsque l'estimation du HC5 généré par cette application est rapportée, d'inclure aussi la version de ssdtools et le nom des fonctions de distribution ajustées aux données.
 
 Les colonnes du tableau de l'évaluation de la qualité de l’ajustement des courbes de distributiont sont la distribution (dist), la statistique d’Anderson-Darling (ad), la statistique de Kolmogorov-Smirnov (ks), la statistique de Cramer-von-Mises (cmv), le critère d’information Akaike (aic), le critère d’information Akaike corrigé pour la taille de l’échantillon (aicc), le critère d’information Bayésien (bic), la différence entre AICc (delta) et la pondération des critères d'information AICc (coefficient de pondération). L’estimation de la fonction de distribution finale est basée sur l’inférence multimodèle (à partir de l’AICc). La concentration présentant un risque est la concentration estimée d’une substance affectant un centile (seuil) sélectionné de l’ensemble des espèces.
 
