### 2020-11-05 #2319 @Kevin-Mattheus-Moerman
@sebdalgarno after further review we decided this work is in scope for JOSS. We will now proceed to find an editor to handle this submission.
