### 2020-11-19 d4a99d7
revisions to paper prior to review statement of need, add ssdtools paper ref and mention SSD Toolbox
