+++ paper/paper.bib
@@ -1,3 +1,21 @@
+@manual{ssdtoolbox_2020, 
+title = {{Species Sensitivity Distribution (SSD) Toolbox}},
+author = {Center for Computational Toxicology and Exposure},
+year = {2020},
+doi = {10.23645/epacomptox.11971392.v1}}
+
+ @article{fox_2020, 
+title = {{Recent developments in SSD Modeling}}, 
+author = {Fox, D.R. and Dam, R.A. and Fisher, R. and Batley, G.E. and Tillmanns, A.R. and Thorley, J. and Schwarz, C.J. and Spry, D.J. and McTavish, K.}, 
+journal = {Environmental Toxicology and Chemistry}, 
+issn = {0730-7268}, 
+doi = {10.1002/etc.4925}, 
+pmid = {33170526}, 
+abstract = {{The species sensitivity distribution (SSD) is a statistical approach that is used to estimate either the concentration of a chemical that is hazardous to no more than x\% of all species (the HCx) or the proportion of species potentially affected by a given concentration of a chemical. Despite a significant body of published research and critical reviews over the past 20 years aimed at improving the methodology, the fundamentals remain unchanged. While there have been some recent suggestions for improvements to SSD methods in the literature, in general, few of these suggestions have been formally adopted. Further, critics of the approach can rightly point to the fact that differences in technical implementation can lead to marked differences in results, thereby undermining confidence in SSD approaches. Despite the limitations, SSDs remain a practical tool and, until a demonstrably better inferential framework is available, developments and enhancements to conventional SSD practice will and should continue. We therefore believe the time has come for the scientific community to decide how it wants SSD methods to evolve. The current paper summarises the current status of, and elaborates on several recent developments for, SSD methods, specifically, model averaging, multimodality and software development. The paper also considers future directions with respect to the use of SSDs, with the ultimate aim of helping to facilitate greater international collaboration and, potentially, greater harmonization of SSD methods. This article is protected by copyright. All rights reserved.}}, 
+local-url = {file://localhost/Users/sebastiandalgarno/Downloads/etc.4925.pdf}, 
+year = {2020}
+}
+
  @book{posthuma_species_2001,
 	title = {Species {Sensitivity} {Distributions} in {Ecotoxicology}},
 	isbn = {978-1-56670-578-3},
