+++ paper/paper.md
@@ -1,6 +1,6 @@
 ---
 title: 'shinyssdtools: A web application for fitting Species Sensitivity Distributions (SSDs)'
-date: '2020-06-02'
+date: '2020-11-18'
 authors:
 - affiliation: 1
   name: Seb Dalgarno
@@ -17,9 +17,14 @@ tags:
 ---
 
 # Summary
-The species sensitivity distribution (SSD) is the most widely used method for getting water quality benchmarks to characterize effects of chemical contaminants for water quality or ecological risk assessment [@fox_2020]. This typically involves estimating the concentration of a chemical that affects a 5% of the species considered [@posthuma_species_2001]. The `ssdtools` R package, developed for the British Columbia Ministry of Environment and Climate Change Strategy [@ssdtools], has recently advanced SSD methods by providing model averaging using information-theoretic criteria and the construction of confidence intervals using bootstrapping [@fox_2020]. 
+The species sensitivity distribution (SSD) is the most widely used method for getting water quality benchmarks to characterize effects of chemical contaminants for water quality or ecological risk assessment [@fox_2020].
+This typically involves estimating the concentration of a chemical that affects 5% of the species considered [@posthuma_species_2001].
+The `ssdtools` R package [@ssdtools] has recently advanced SSD methods by providing model averaging using information-theoretic criteria and the construction of confidence intervals using bootstrapping [@fox_2020]. 
 
-`shinyssdtools` is a web-based graphical user interface (GUI) and R package built with the Shiny web framework [@shiny]. It provides access to the core functionality in the `ssdtools` R package via a simple and modern user interface. It also provides the following functionality in addition to `ssdtools`: translation to French; generation of R code to reproduce results from a session; further customization of plot and table outputs; download of tables and plot outputs. `shinyssdtools` can be accessed from the web or it can be run locally by installing the R package. The advantage of using a web application over standalone software is that it is possible to ensure that the user is using the most up-to-date version. The advantage of developing the application as an R package is that the source code is openly visible and is available for local modification. 
+`shinyssdtools` is a web-based graphical user interface (GUI) to the `ssdtools` R package.
+`shinyssdtools`, which was developed in the Shiny web framework [@shiny], is an R package in its own right.
+As well as providing access to the core functionality in `ssdtools` it also offers the following value-added features: a bilingual (English/French) interface, customization and downloads of plots and tables and generation of R scripts for reproducibility. 
+`shinyssdtools` can be accessed from the web or it can be run locally by installing the R package.
 
 ### Graphical User Interface
 
@@ -54,16 +59,25 @@ shinyssdtools::run_ssdtools_app()
 ```
 
 # Statement of Need
-`ssdtools` is an R package that has provided recent advances in SSD methodology [@fox_2020]. `shinyssdtools` provides access to this functionality via a simple, modern GUI without requiring users to be familiar with the R programming language. Data can be easily uploaded to the application, the interface can be viewed in multiple languages (French and English) and R code output is provided so that analyses can be reproduced or shared. `shinyssdtools` was developed for the Province of British Columbia with input from the governments of Canada as well as Australia and New Zealand and has been used by the governments of B.C. and Canada to derive water quality benchmarks.
+`ssdtools` is an R package that has provided recent advances in SSD methodology [@fox_2020]. 
+`shinyssdtools` provides access to this functionality via a simple, modern GUI without requiring users to be familiar with the R programming language. 
+Data can be easily uploaded to the application, the interface can be viewed in multiple languages (French and English) and R code output is provided so that analyses can be reproduced or shared. 
[…]
