+++ paper/paper.bib
@@ -1,8 +1,10 @@
 @manual{ssdtoolbox_2020, 
-title = {{Species Sensitivity Distribution (SSD) Toolbox}},
-author = {Center for Computational Toxicology and Exposure},
-year = {2020},
-doi = {10.23645/epacomptox.11971392.v1}}
+title = {{User’s Manual: SSD Toolbox Version 1.0}},
+author = {Etterson, M.},
+organization = {US Environmental Protection Agency, Office of Research and Development, Center for Computational Toxicology and Exposure},
+year = {2020}}
+
+Etterson M. 2020. User’s Manual: SSD Toolbox Version 1.0. US Environmental Protection Agency, Office of Research and Development, Center for Computational Toxicology and Exposure.
 
  @article{fox_2020, 
 title = {{Recent developments in SSD Modeling}}, 
