### 2020-11-19 #2319 @sebdalgarno
> Hi, @sebdalgarno, I will be the editor of your submission.
>
> I see you already sent one reviewer suggestion (meenakshi-kushwaha), and I'll start looking for two more - let me know if you have any other potential reviewers to suggest, ok? If you haven't see it yet, we have [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS.
>
> In the meantime, please include the "statement of need" a separate paragraph, it should look like [this](https://joss.theoj.org/papers/10.21105/joss.02584) or [this](https://joss.theoj.org/papers/10.21105/joss.02342).
>
> Also, do you think you can expand the explanation on what does the package do with a more general, less technical explanation? JOSS guidelines for the summary asks for a "description of the high-level functionality and purpose of the software for a diverse, non-specialist audience". Usually, this can be easily achieved with one or two phrases at the beggening of the summary (and I think that both papers linked above are good examples for that).

Thanks @marcosvital ! I've made the requested revisions (statement of need and some added context at beginning). I've also added reference to a recently published article on recent advances in SSD methods and added a section to the contributions section comparing shinyssdtools to the US EPA ssd Toolbox
