### 2020-11-19 #2319 @marcosvital
Thanks for the revisions, @sebdalgarno. The proof seems ready for revision.

@nanhung, thank you. We will start the revision process right away.
