+++ paper/paper.bib
@@ -4,8 +4,6 @@ author = {Etterson, M.},
 organization = {US Environmental Protection Agency, Office of Research and Development, Center for Computational Toxicology and Exposure},
 year = {2020}}
 
-Etterson M. 2020. User’s Manual: SSD Toolbox Version 1.0. US Environmental Protection Agency, Office of Research and Development, Center for Computational Toxicology and Exposure.
-
  @article{fox_2020, 
 title = {{Recent developments in SSD Modeling}}, 
 author = {Fox, D.R. and Dam, R.A. and Fisher, R. and Batley, G.E. and Tillmanns, A.R. and Thorley, J. and Schwarz, C.J. and Spry, D.J. and McTavish, K.}, 
@@ -22,6 +20,7 @@ year = {2020}
 	title = {Species {Sensitivity} {Distributions} in {Ecotoxicology}},
 	isbn = {978-1-56670-578-3},
 	publisher = {CRC press},
+	doi = {10.1201/9781420032314},
 	author = {Posthuma, Leo and Suter II, Glenn W and Traas, Theo P},
 	year = {2001}
 }
