### 2020-11-26 bcgov/shinyssdtools#58 @elimillera
JOSS Review: Licence Template and Metadata Updates
Per the JOSS review checklist:

> License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

1. The Apache 2.0 license is included however the placeholders for the copyright holder and year haven't been updated.
Line 189 in the LICENSE file:
`Copyright {yyyy} {name of copyright owner}`

2. The DESCRIPTION file in the top-level directory lists the file as MIT. If this is incorrect you can replace it with 'Apache License | file LICENSE'
Line 9 in the DESCRIPTION file:
`License: MIT + file LICENSE`

3. The copyright holder doesn't appear in the list of "Authors" in the DESCRIPTION file. While this isn't a JOSS issue this would prevent [submission to CRAN](https://cran.r-project.org/web/packages/policies.html#Source-packages) if you are planning on that. [Here](https://github.com/tidyverse/dplyr/blob/master/DESCRIPTION) is how dplyr lists RStudio as the copyright holder.
Line 4-7 in the DESCRIPTION file:
```
Authors@R: c(
    person("Joe", "Thorley", , "joe@poissonconsulting.ca", "ctb", comment = c(ORCID = "0000-0002-7683-4592")),
    person("Sebastian", "Dalgarno", , "seb@poissonconsulting.ca", c("aut", "cre"), comment = c(ORCID = "0000-0002-3658-4517"))
    )
```
