### 2020-11-26 bcgov/shinyssdtools#60 @elimillera
JOSS Review: Title of Shiny Application is Raw HTML
When running the app the title of the browser is an HTML tag instead of the intended title. I tried playing around with the "ui_navtitle" output but I'm not sure how to make the title reactive.

![image](https://user-images.githubusercontent.com/24767886/100393898-db299200-3000-11eb-925f-bd550cbad15a.png)
