### 2020-11-26 #2848 @elimillera
@sebdalgarno My review is complete. Should be good to go once the [license issues](https://github.com/bcgov/shinyssdtools/issues/58) are resolved.
