### 2020-11-28 #2848 @nanhung
@sebdalgarno I appreciate your developed shiny app that makes **ssdtools** more accessible and easy to use. I like this "minimalist" design, but have some comments to you to improve your app. Here are my comments after testing **shinyssdtools**.

My main recommendation is about the interface. I suggest you can use different colors, font types (e.g., bold), and size to make the interface more friendly to your users. For example in **`1. Data`**, The title `Choose one of the following options:` should be emphasized by using a larger font or bold. Also, the font of Note on the bottom can be emphasized as well.

## Some specific comments are:

1. Starting from **`1. Data`**. Is it possible to add a widget to define the number of rows (or columns) that the users want to add? So they do not need to right-click the table and add a single row (or column), repeatedly. In addition, I find that the column name is not able to revise. Can you add the function to edit them? Moreover, can you add a function that can let users export the CSV file that they created in the table?

2. When finishing up the data import. The current setting is to move the cursor and click the **`2.Fit`** tab on the top ribbon. I would like it if you can add the **[Next]** button at the panel bottom. I think it is very intuitive to general users. It takes me a while to know how this app work.

3. When using the  Boron data in the model fitting. It shows "llogis, lgumbel, gompertz distribution(s) failed to fit. Run R code to get more information."

![Screenshot 2020-11-27 160449](https://user-images.githubusercontent.com/25087463/100489652-57d47300-30ca-11eb-82ca-b55b5e8a1d84.jpg)

However, these models work in the **ssdtools**.

![Screenshot 2020-11-27 160321](https://user-images.githubusercontent.com/25087463/100489620-1fcd3000-30ca-11eb-8938-e3f04f91be7a.jpg)

You might need to check this bug. Also, Suggest using the different font to emphasize the warning message.

4. In **`3. Predicted`**. The same comment to enhance the title `Estimate hazard concentration` using a larger font.

5. I don't understand the usage of `plot.rds`?

6. Another feature request: Add CL in the predicted plot.

7. The dashed line (threshold) does not appear when using the concentration as the threshold.

## Minor comments:

1. The current Tab name looks like the source code. Suggest to revise it to the title such as `shinyssdtools v1.0`

![Screenshot 2020-11-27 160834](https://user-images.githubusercontent.com/25087463/100489717-e2b56d80-30ca-11eb-80c8-97a36685028e.jpg)

2. Can you add a **<default>** button to return to the default setting? It is like recovery mode if users want to reset theirs definitions.

3. The weird direction of tabs appeared when adjusting the window size to the mobile mode

![Screenshot 2020-11-27 161319](https://user-images.githubusercontent.com/25087463/100489809-8141ce80-30cb-11eb-9de1-9dbb8c33aa0b.jpg)
