### 2020-12-14 #2848 @meenakshi-kushwaha
Thank you for the reminder. I plan to work on the  the review this week.

> On 12-Dec-2020, at 6:20 AM, Marcos Vital (LEQ-UFAL) <notifications@github.com> wrote:
>
>
> Hi, @meenakshi-kushwaha <https://github.com/meenakshi-kushwaha>, any news about this review? Let us know if you have any problems.
>
> @sebdalgarno <https://github.com/sebdalgarno>, let us know if you made any progress on the issues raised by nanhung, ok?
>
