### 2020-12-21 #2848 @meenakshi-kushwaha
Thank you for inviting me for the review. I think this app is useful in extending the functionalities of the original ssdtools package. I especially like the R code feature that can help the user customize the analysis and plots further. Since authors have already responded to key suggestions from earlier reviewers, I have only a few suggestions left

Minor
- The clean layout app is great.  Authors could consider  using larger font or bold font for table/plot headings - ex. Goodness of Fit table, Plot fitted distributions, etc. for better clarity
- In goodness of fit table title row, consider changing "dist" to Distribution for better aesthetics
- The selected values in the png file formatting options are not completely visible. In the following image,  I have selected 10 and 6
![image](https://user-images.githubusercontent.com/52627451/102791610-38d3b300-43e2-11eb-9796-f933f1067cc4.png)

Major
- Appropriate license file is missing. The current license file in the repo has details of the Apache license, but not the actual license with Copyright information
- I noticed that the project has multiple contributors who are not included as authors in the paper. This is my first JOSS review so I will defer to @marcosvital on expected authorship norms.
![image](https://user-images.githubusercontent.com/52627451/102792256-39b91480-43e3-11eb-80aa-dd12f94afa72.png)

~Meenakshi
