### 2020-12-29 435ee79
fix license issues - fixes #58
