### 2020-12-29 #2848 @sebdalgarno
@elimillera I've made the requested license changes in this commit: https://github.com/sebdalgarno/shinyssdtools/commit/435ee79835584876eac5fc82c46607797335bdfb

I will make a pull request to finalize once I'm done with the rest of the changes.
