### 2020-12-29 bcgov/shinyssdtools#60 @sebdalgarno
fixed by adding tags$head(tags$title(...
