### 2020-12-29 918fa64
make headings bolder fixes #62
