### 2020-12-29 21009f1
add CL to pred plot fixes #63
