### 2020-12-29 #2848 @sebdalgarno
@nanhung I will respond to each of your comments in bold
> @sebdalgarno I appreciate your developed shiny app that makes **ssdtools** more accessible and easy to use. I like this "minimalist" design, but have some comments to you to improve your app. Here are my comments after testing **shinyssdtools**.
>
> My main recommendation is about the interface. I suggest you can use different colors, font types (e.g., bold), and size to make the interface more friendly to your users. For example in **`1. Data`**, The title `Choose one of the following options:` should be emphasized by using a larger font or bold. Also, the font of Note on the bottom can be emphasized as well.
>
> ## Some specific comments are:
> 1. Starting from **`1. Data`**. Is it possible to add a widget to define the number of rows (or columns) that the users want to add? So they do not need to right-click the table and add a single row (or column), repeatedly. In addition, I find that the column name is not able to revise. Can you add the function to edit them? Moreover, can you add a function that can let users export the CSV file that they created in the table?

**This is not possible as it uses the rhandsontable R package which itself has limited features.**

> 2. When finishing up the data import. The current setting is to move the cursor and click the **`2.Fit`** tab on the top ribbon. I would like it if you can add the **[Next]** button at the panel bottom. I think it is very intuitive to general users. It takes me a while to know how this app work.

**I don't agree that this would help as I think that numbering the tabs makes this obvious**

> 3. When using the  Boron data in the model fitting. It shows "llogis, lgumbel, gompertz distribution(s) failed to fit. Run R code to get more information."
>
> ![Screenshot 2020-11-27 160449](https://user-images.githubusercontent.com/25087463/100489652-57d47300-30ca-11eb-82ca-b55b5e8a1d84.jpg)
>
> However, these models work in the **ssdtools**.
>
> ![Screenshot 2020-11-27 160321](https://user-images.githubusercontent.com/25087463/100489620-1fcd3000-30ca-11eb-8938-e3f04f91be7a.jpg)
>
> You might need to check this bug. Also, Suggest using the different font to emphasize the warning message.

**I cannot reproduce this issue. Perhaps you need to reinstall the package?**

>
> 1. In **`3. Predicted`**. The same comment to enhance the title `Estimate hazard concentration` using a larger font.

**fixed this in https://github.com/bcgov/shinyssdtools/issues/62**

> 2. I don't understand the usage of `plot.rds`?

**this is useful for R users to download the ggpot2 object and work with it**

> 3. Another feature request: Add CL in the predicted plot.

**fixed in this issue: https://github.com/bcgov/shinyssdtools/issues/63**

> 4. The dashed line (threshold) does not appear when using the concentration as the threshold.

**This is not important enough to warrant the required changes that would have to be made to the ssdtools R package (which I don't maintain)**

>
> ## Minor comments:
> 1. The current Tab name looks like the source code. Suggest to revise it to the title such as `shinyssdtools v1.0`

**fixed in this issue: https://github.com/bcgov/shinyssdtools/issues/60**

>
> ![Screenshot 2020-11-27 160834](https://user-images.githubusercontent.com/25087463/100489717-e2b56d80-30ca-11eb-80c8-97a36685028e.jpg)
>
> 1. Can you add a  button to return to the default setting? It is like recovery mode if users want to reset theirs definitions.

**I don't agree that this is necessary**

> 2. The weird direction of tabs appeared when adjusting the window size to the mobile mode

**This is due to bootstrap. There is no way to change this within my chosen framework**

>
> ![Screenshot 2020-11-27 161319](https://user-images.githubusercontent.com/25087463/100489809-8141ce80-30cb-11eb-9de1-9dbb8c33aa0b.jpg)
