### 2020-12-29 bcgov/shinyssdtools#64 @sebdalgarno
JOSS Review: png formatting numeric inputs too small
can fix by not making inline
