### 2020-12-29 bcgov/shinyssdtools#58 @sebdalgarno
thanks @elimillera for the helpful and detailed comments. Should be all good now
