### 2020-12-29 #2848 @sebdalgarno
@marcosvital everything has been addressed. Please see issues that were opened in reference to each comment/suggestion/problem and associated commits.

Changes have been pushed to this fork and branch: https://github.com/sebdalgarno/shinyssdtools/tree/joss_review

Next step is to get Province of BC to merge a pull request and redeploy the live site. However, it is possible for the reviewers to see the changes I've made by installing from the fork, e.g. `remotes::install_github("sebdalgarno/shinyssdtools@joss_review")`
