### 2021-01-06 bcgov/shinyssdtools#58 @elimillera
@sebdalgarno Are you planning on submitting into CRAN? Item three above doesn't look to be resolved. This is an issue specific to CRAN policies and I don't think this should hold up anything with JOSS. However, I have had a package get rejected from CRAN for this reason.

Thanks for the updates!
