### 2021-01-06 bcgov/shinyssdtools#58 @sebdalgarno
@elimillera no plans to submit to CRAN right now. But will keep this in mind if Province of BC requests that it be submitted to CRAN. Thanks!
