### 2021-01-06 #2848 @sebdalgarno
@marcosvital my employer requested a slight change to affiliation ('Poisson Consulting' should be 'Poisson Consulting Ltd.'). I have made the change on a PR. Can I merge the PR and rebuild the paper?
