### 2021-01-15 #2848 @marcosvital
Hi, @sebdalgarno, sorry about the delayed reply. I believe we are almost ready to go here. For now, I would like you to:

* make a final check the latest version of the manuscript generated by Whedon, to see if everything is ok.
* archive the last release of the package (on Zenodo, figshare, or other) if you already didn't do this. After that, just let me know the archive DOI.
