### 2021-01-18 b8908ff
bump version - major release with JOSS paper
