### 2021-01-18 5477e0f
remove deprecated butler funs
