### 2021-01-25 #2848 @danielskatz
@sebdalgarno  - I've suggested two small changes in https://github.com/bcgov/shinyssdtools/pull/68 - please merge this when you get a chance
